requirements.txt
langchain
langchain_community
langchain-openai
streamlit
openai
python-dotenv